+++ docker-entrypoint.sh
@@ -2,8 +2,9 @@
 
 set -e
 
-service redis-server start
-sleep 1
-service redis-server status # returns non-0 if not running
+# Update 2020-01-02 Dropped redis server from docker image
+# service redis-server start
+# sleep 1
+# service redis-server status # returns non-0 if not running
 
 python3 -m pew in isitfit "$@"